LC-maximum-area-of-longest-diagonal-rectangle [Runtime: N/A; Memory: N/A]

+++ array/maximum-area-of-longest-diagonal-rectangle.cpp
@@ -2,12 +2,12 @@ class Solution {
 public:
     int areaOfMaxDiagonal(vector<vector<int>>& dimensions) {
         int l,b,area;
-        int diagonal=0;
+        float diagonal=0;
 
         for(auto it: dimensions){
-            int li=it[0];
-            int bi=it[1];
-            int root=sqrt((li*li) + (bi*bi));
+            float li=it[0];
+            float bi=it[1];
+            float root=sqrt((li*li) + (bi*bi));
             if(root>diagonal){
                 diagonal=root;
                 l=li;
